Refactor parameter sets
