+++ setup.cfg
@@ -75,10 +75,10 @@ ewatercycle.forcings =
     MarrmotForcing = ewatercycle.plugins.marrmot.forcing:MarrmotForcing
     PCRGlobWBForcing = ewatercycle.plugins.pcrglobwb.forcing:PCRGlobWBForcing
     WflowForcing = ewatercycle.plugins.wflow.forcing:WflowForcing
-ewatercycle.example_parameter_set = 
-    lisflood_fraser = ewatercycle.plugins.lisflood.example_parameterset:lisflood_fraser
-    pcrglobwb_rhinemeuse_30min = ewatercycle.plugins.pcrglobwb.example_parameterset:pcrglobwb_rhinemeuse_30min
-    wflow = ewatercycle.plugins.wflow:example_parameterset
+ewatercycle.parameter_sets = 
+    lisflood_fraser = ewatercycle.plugins.lisflood.parameter_sets:lisflood_fraser
+    pcrglobwb_rhinemeuse_30min = ewatercycle.plugins.pcrglobwb.parameter_sets:pcrglobwb_rhinemeuse_30min
+    wflow_rhine_sbm_nc = ewatercycle.plugins.wflow.parameter_sets:wflow_rhine_sbm_nc
 
 [options.extras_require]
 dev =
