+++ src/ewatercycle/plugins/lisflood/model.py
@@ -12,7 +12,7 @@ from cftime import num2date
 from ewatercycle import CFG
 from ewatercycle.container import start_container
 from ewatercycle.model import AbstractModel
-from ewatercycle.parameter_sets import ParameterSet
+from ewatercycle.base.parameter_set import ParameterSet
 from ewatercycle.parametersetdb.config import XmlConfig
 from ewatercycle.plugins.lisflood._lisflood_versions import version_images
 from ewatercycle.plugins.lisflood.forcing import LisfloodForcing
