+++ src/ewatercycle/plugins/lisflood/parameter_sets.py
@@ -1,10 +1,10 @@
-from ewatercycle.parameter_set import ParameterSet
+from ewatercycle.base.parameter_set import ParameterSet
 
 
 lisflood_fraser = ParameterSet.from_github(
     repo='https://github.com/ec-jrc/lisflood-usecases/tree/master/LF_lat_lon_UseCase',
     name="lisflood_fraser",
-    # Relative to CFG.parameterset_dir / self.name
+    directory="lisflood_fraser",
     config="settings_lat_lon-Run.xml",
     doi="N/A",
     target_model="lisflood",
