+++ src/ewatercycle/plugins/wflow/parameter_sets.py
@@ -1,9 +1,9 @@
-from ewatercycle.parameter_set import ParameterSet
+from ewatercycle.base.parameter_set import ParameterSet
 
 wflow_rhine_sbm_nc = ParameterSet.from_github(
     repo='https://github.com/openstreams/wflow/tree/master/examples/wflow_rhine_sbm_nc',
+    directory="wflow_rhine_sbm_nc",
     name="wflow_rhine_sbm_nc",
-    # Relative to CFG.parameterset_dir / self.name
     config="wflow_sbm_NC.ini",
     doi="N/A",
     target_model="wflow",
